pyproject.toml
[project]
name = "pixel-magic"
version = "0.1.0"
description = "AI-powered pixel art sprite generation and conversion pipeline with MCP server"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    "google-genai>=1.0.0",
    "openai>=1.60.0",
    "mcp[cli]>=1.20.0",
    "opencv-python-headless>=4.9.0",
    "Pillow>=10.0.0",
    "scikit-image>=0.22.0",
    "numpy>=1.26.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
pixel-magic = "pixel_magic.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pixel_magic"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
